LayoutTests/editing/pasteboard/drag-drop-href-as-text-data-expected.txt


text: "https://www.webkit.org/"
url: "https://www.webkit.org/"
text/plain: "https://www.webkit.org/"
text/uri-list: "https://www.webkit.org/"
